[Chore] 구조체 위치 변경, 디버깅 로직 추가
